requirements.txt
asgiref==3.2.10
Django==2.2
django-appconf==1.0.4
django-crispy-forms==1.9.2
django-imagekit==4.0.2
django-taggit==1.3.0
pilkit==2.0
Pillow==7.2.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
